requirements.txt
fastapi
uvicorn
numpy
pillow
python-multipart
torch
transformers